+++ mockserver-core/src/main/java/org/mockserver/collections/NottableStringMultiMap.java
@@ -2,6 +2,7 @@ package org.mockserver.collections;
 
 import com.google.common.annotations.VisibleForTesting;
 import org.mockserver.logging.MockServerLogger;
+import org.mockserver.matchers.MatchDifference;
 import org.mockserver.matchers.RegexStringMatcher;
 import org.mockserver.model.KeyMatchStyle;
 import org.mockserver.model.KeyToMultiValue;
@@ -41,15 +42,26 @@ public class NottableStringMultiMap extends ObjectWithReflectiveEqualsHashCodeTo
         }
     }
 
-    public boolean containsAll(NottableStringMultiMap subset) {
+    public KeyMatchStyle getKeyMatchStyle() {
+        return keyMatchStyle;
+    }
+
+    public boolean containsAll(MockServerLogger mockServerLogger, MatchDifference context, NottableStringMultiMap subset) {
         switch (subset.keyMatchStyle) {
             case SUB_SET: {
-                return containsSubset(regexStringMatcher, subset.entryList(), entryList());
+                boolean isSubset = containsSubset(regexStringMatcher, subset.entryList(), entryList());
+                if (!isSubset && context != null) {
+                    context.addDifference(mockServerLogger, "multimap subset match failed subset:{}was not a subset of:{}", subset.entryList(), entryList());
+                }
+                return isSubset;
             }
             case MATCHING_KEY: {
                 for (NottableString matcherKey : subset.backingMap.keySet()) {
                     List<NottableString> matchedValuesForKey = getAll(matcherKey);
                     if (matchedValuesForKey.isEmpty() && !matcherKey.isOptional()) {
+                        if (context != null) {
+                            context.addDifference(mockServerLogger, "multimap subset match failed subset:{}did not have expected key:{}", subset, matcherKey);
+                        }
                         return false;
                     }
 
@@ -57,9 +69,13 @@ public class NottableStringMultiMap extends ObjectWithReflectiveEqualsHashCodeTo
                     for (NottableString matchedValue : matchedValuesForKey) {
                         boolean matchesValue = false;
                         for (NottableString matcherValue : matcherValuesForKey) {
-                            if (regexStringMatcher.matches(matcherValue, matchedValue)) {
+                            if (regexStringMatcher.matches(mockServerLogger, context, matcherValue, matchedValue)) {
                                 matchesValue = true;
                                 break;
+                            } else {
+                                if (context != null) {
+                                    context.addDifference(mockServerLogger, "multimap matching key match failed for key:{}", matcherKey);
+                                }
                             }
                         }
                         if (!matchesValue) {